chiusi i punti 1 e 2 della verifica: render_preview legge i token dal CSS e ruff a 0
§7.1 — scripts/render_preview.py: niente più copie a mano.
- I 24 token del tema scuro, il gradiente dell'header e i colori del badge si leggono
  a runtime da assets/site.css e templates/base.html (verificato 24/24 uguali al CSS;
  se un token sparisce lo script si ferma nominandolo). I colori V/N/P non sono più
  approssimazioni ma i token veri.
- PAGE e CARDS si leggono da site/index.html: l'anteprima segue il build invece di
  invecchiare. Senza build (CI) usa uno snapshot di riserva con i valori veri del
  18/09 e lo dichiara a schermo.
- Rigenerati docs/preview/*.png con la palette corrente.
- Test nuovi: colori letti dal CSS (fissa il collegamento) e parser della home
  verificato su markup sintetico (se un campo della card torna vuoto, il test lo dice).

§7.2 — ruff da 172 a 0.
- 135 correzioni sicure (--fix) + 28 meccaniche (--unsafe-fixes, verificate coi
  test) + chiusure manuali: 6 B023 legati via default, 3 SIM102, 3 RUF012 -> ClassVar,
  3 F401, RUF034, UP031/032, I001.
- Le categorie intenzionali (BLE001/S110 nei collettori "degrada e pubblica comunque",
  B008 = idioma typer, DTZ001 in un test con date naive) sono eccezioni per-file
  documentate in pyproject [tool.ruff.lint.per-file-ignores]: restano vietate nei file
  nuovi (regola B.4) e il gate torna verde. Non eseguito ruff format (non è nel gate).

Gate finale: pytest 433 passed · ruff 0 · fda build exit 0 · verify_site exit 0
(97.872 controlli + 186.054 attributi). Report aggiornato in docs/27 §7.

+++ tests/test_backtest.py
@@ -134,7 +134,7 @@ def test_backtest_summary_metrics_are_recomputed_independently():
     assert all(c["n"] == n for c in s["calib"])
     # Over 1,5 dichiarato 75% su esiti 2-0/1-1/0-2: osservato ben sotto → fuori intervallo
     o15 = next(m for m in s["markets"] if m["label"] == "Over 1,5 gol")
-    lo, hi = wilson_interval(int(round(o15["obs"] * o15["n"])), o15["n"])
+    lo, hi = wilson_interval(round(o15["obs"] * o15["n"]), o15["n"])
     assert o15["outside"] is bool(not (lo <= o15["prev"] <= hi))
     for m in s["markets"]:
         assert abs(m["delta"] - (m["brier"] - m["brier_base"])) < 1e-12
